# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Property Lo API — Nigerian real estate listing platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: purpose
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
    post:
      operationId: createListing
      tags: [listings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    patch:
      operationId: updateListing
      tags: [listings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Favorites ─────────────────────────────────────────────────────────────
...
  /support/chat:
    get:
      operationId: getSupportChat
      tags: [support]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Support chat messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportChatResponse"
  /support/chat/messages:
    post:
      operationId: sendSupportMessage
      tags: [support]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"
  /support/unread-count:
    get:
      operationId: getSupportUnreadCount
      tags: [support]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCountResponse"

  # ── Ads ───────────────────────────────────────────────────────────────────
...
  /admin/listings:
    get:
      operationId: getAdminListings
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    UnreadCountResponse:
      type: object
      required: [count]
      properties:
        count:
          type: integer
...
    Listing:
      type: object
      required: [id, title, description, type, purpose, price, location, state, images, status, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
        purpose:
          type: string
        price:
          type: number
        location:
          type: string
        state:
          type: string
        size:
          type: ["string", "null"]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
        contact_email:
          type: ["string", "null"]
        contact_phone:
          type: ["string", "null"]
        user_id:
          type: ["string", "null"]
        created_at:
          type: string
    ListingsResponse:
      type: object
      required: [listings]
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
...
    ListingInput:
      type: object
      required: [title, description, type, purpose, price, location, state]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        purpose:
          type: string
        price:
          type: number
        location:
          type: string
        state:
          type: string
        size:
          type: string
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        images:
          type: array
          items:
            type: string
        status:
          type: string
        contact_email:
          type: string
        contact_phone:
          type: string
    ListingUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        purpose:
          type: string
        price:
          type: number
        location:
          type: string
        state:
          type: string
        size:
          type: string
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        images:
          type: array
          items:
            type: string
        status:
          type: string
        contact_email:
          type: string
        contact_phone:
          type: string

    # ── Favorites ─────────────────────────────────────────────────────────────
...
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Support ───────────────────────────────────────────────────────────────
    SupportMessage:
      type: object
      required: [id, content, is_from_user, created_at]
      properties:
        id:
          type: string
        content:
          type: string
        is_from_user:
          type: boolean
        created_at:
          type: string
    SupportChatResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportMessage"

    # ── Ads ───────────────────────────────────────────────────────────────────
...